# Introduce accurate formulas for determinant and inverse of small matrices