xtensor and numpy compute the determinant using LU decomposition.
This PR introduces formulas that are more accurate and faster for matrices up to 3x3.

Impact on performance:
```python3
mesh = dolfinx.UnitCubeMesh(comm, 100, 100, 100, CellType.tetrahedron)
V = dolfinx.FunctionSpace(mesh, ("N1curl", 1))
```

| | Main | This branch |
| :---         |     :---:      |          :---: |
| `compute_point_values`   | 21.30 s    | 17.13 s   |
| `interpolate`    | 47.74 s     | 32.91 s    |

```python3
mesh = dolfinx.UnitCubeMesh(comm, 100, 100, 100, CellType.hexahedron)
V = dolfinx.FunctionSpace(mesh, ("NCE", 1))
```
| | Main | This branch |
| :---         |     :---:      |          :---: |
| `compute_point_values`   | 19.59 s| 15.64 s  |
| `interpolate`    |  58.31 s      | 16.42 s  |


```python3
mesh = dolfinx.UnitCubeMesh(comm, 100, 100, 100, CellType.Tetrahedron)
V = dolfinx.FunctionSpace(mesh, ("Lagrange", 1))
```
| | Main | This branch |
| :---         |     :---:      |          :---: |
| `compute_point_values`   |  25.09  s   | 13.03 s  |
| `interpolate`    |  10.69  s    | 10.70 s |