@garth-wells 2021-04-18
I think the namespace `linalg` is potentially confusing (confusing with `xt::linalg`, and confusing with `dolfinx::la`). Maybe `math` would be better?